Refactor add, edit, delete deck
